+++ docs/changelog.rst
@@ -12,6 +12,7 @@ New:
   database changes. Thanks again to xsteadfastx. :bug:`1120`
 * Add a warning when importing a directory that contains no music. :bug:`1116`
   :bug:`1127`
+* :doc:`/plugins/zero` can remove embedded images.
 
 Fixed: